+++ DataAccess/Models/CustomersModel.cs
@@ -1,11 +1,11 @@
 ﻿namespace DataAccess.Models;
 public class CustomersModel
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string? Name { get; set; }
     public string? Email { get; set; }
     public string? Address { get; set; }
-    public int PriceGroupId { get; set; }
+    public long PriceGroupId { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
 }
