+++ DataAccess/Models/OrdersModel.cs
@@ -1,8 +1,8 @@
 ﻿namespace DataAccess.Models;
 public class OrdersModel
 {
-    public int Id { get; set; }
-    public int CustomerId { get; set; }
+    public long Id { get; set; }
+    public long CustomerId { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
 }
